refactor: update config resolver and feature type definitions
