--- src/utils/index.ts
@@ -1,4 +1,0 @@
-export * from "./config.js";
-export * from "./config-loader.js";
-export * from "./error.js";
-export * from "./file.js";
